+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Unreleased
 
-## Bug fixes
+ * Drop ezjsonm dependency for `gitlab` (#76 @tmcgilchrist)
 
+## Bug fixes
  * `gitlab.atd`: `description` is nullable (#75 @maiste)
 
 # 0.1.6 - 2022-11-08